feat: menu refactored
