+++ packages/anu-vue/src/components/menu/index.ts
@@ -1,3 +1,3 @@
-export { AMenu } from './AMenu'
+export { default as AMenu } from './AMenu.vue'
 export { sameWidth as sameWidthFloatingUIMiddleware } from './middlewares'
